relocate manifest and compose functions
