+++ src/releaseFiles/compose/compose.ts
@@ -2,18 +2,18 @@ import fs from "fs";
 import path from "path";
 import yaml from "js-yaml";
 import prettier from "prettier";
-import { PackageImage } from "../types";
+import { PackageImage } from "../../types";
 import {
   defaultComposeFileName,
   defaultDir,
   getImageTag,
   upstreamImageLabel,
   UPSTREAM_VERSION_VARNAME
-} from "../params";
-import { toTitleCase } from "./format";
+} from "../../params";
+import { toTitleCase } from "../../utils/format";
 import { mapValues, uniqBy } from "lodash";
-import { readFile } from "./file";
-import { Compose } from "../releaseFiles/compose/types";
+import { readFile } from "../../utils/file";
+import { Compose } from "./types";
 
 interface ComposePaths {
   /** './folder', [optional] directory to load the compose from */
