+++ src/releaseFiles/manifest/manifest.ts
@@ -2,9 +2,9 @@ import fs from "fs";
 import path from "path";
 import yaml from "js-yaml";
 import prettier from "prettier";
-import { defaultDir, releaseFiles } from "../params";
-import { readFile } from "./file";
-import { Manifest, ManifestFormat } from "../releaseFiles/manifest/types";
+import { defaultDir, releaseFiles } from "../../params";
+import { readFile } from "../../utils/file";
+import { Manifest, ManifestFormat } from "./types";
 
 export interface ManifestPaths {
   /** './folder', [optional] directory to load the manifest from */
